// demo-ui/src/app/features/generated/organizations/organization/models/OrganizationFullView.ts
import { OrganizationForm } from './OrganizationForm';
/**
 * auto generated full veiw bean ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class OrganizationFullView {

id: number;
form: OrganizationForm;
canUpdate: boolean;

/* Specific Code Start */
/* Specific Code End */
}